+++ server/src/entity/Order.cpp
@@ -21,11 +21,11 @@ void Order::setUid(size_t uid) {
     Order::uid = uid;
 }
 
-const float Order::getCost() const {
+const double Order::getCost() const {
     return cost;
 }
 
-void Order::setCost(float cost) {
+void Order::setCost(double cost) {
     Order::cost = cost;
 }
 
@@ -69,7 +69,7 @@ void from_json(const nlohmann::json &j, Order &o) {
 
 size_t Order::id_count = 0;
 
-Order::Order(size_t uid,float cost, time_t start, time_t end, ChargingType type)
+Order::Order(size_t uid,double cost, time_t start, time_t end, ChargingType type)
 : id(id_count++),cost(cost), uid(uid), start(start), end(end), type(type){
 
 }
